Add provisional defanging algorithm
There's a few different cases required in order for the algorithm to
pass the safety checks, but I think this is probably a pretty good
first pass at such an algorithm to extend defang support for other
schemes.  The most fragile of rules are in the case where the length
of the scheme is 2 letters.  But for the moment, this does not cause
any issues.